mais um dia

+++ routes.py
@@ -1,6 +1,6 @@
 from app import app,db, login_manager
 from flask import render_template, redirect, url_for, flash
-from models import UserModel, UserPostModel
+from models import UserModel, UserPostModel, UserEnderecoModel
 from forms import UserForm, LoginForm, UserPostForm, SearchForm, AlterUserForm, AlterUserPasswordForm
 from flask_login import login_required, logout_user, current_user, login_user
 from werkzeug.security import generate_password_hash, check_password_hash
@@ -35,27 +35,46 @@ def index():
     cozinheiros = UserModel.query.filter(UserModel.status == True).all()
     return render_template('index.html',posts=posts, cozinheiros=cozinheiros)
 
-@app.route('/add_user',methods=['POST','GET'])
+@app.route('/add_user', methods=['POST', 'GET'])
 def add_user():
     form = UserForm()
     if form.validate_on_submit():
         foto = form.profile_pic.data
-        profile_pic = "{}_{}".format(uuid1(),secure_filename(foto.filename))
+        profile_pic = None
+        if foto:
+            profile_pic = "{}_{}".format(uuid1(), secure_filename(foto.filename))
+            try:
+                foto.save(os.path.join(app.config['UPLOAD_FOLDER'], profile_pic))
+            except Exception as e:
+                flash('Error uploading the picture')
+                return redirect(url_for('add_user'))
+
         user = UserModel(
-            name = form.name.data,
-            email = form.email.data,
-            username = form.username.data,
-            password_hash = generate_password_hash(form.password.data, method='pbkdf2:sha256'),
-            profile_pic = profile_pic
+            name=form.name.data,
+            email=form.email.data,
+            username=form.username.data,
+            password_hash=generate_password_hash(form.password.data, method='pbkdf2:sha256'),
+            profile_pic=profile_pic,
+            cozinheiro=bool(form.cozinheiro.data)
         )
-        foto.save(os.path.join(app.config['UPLOAD_FOLDER'], profile_pic))
-        form = UserForm(formdata=None)
         db.session.add(user)
         db.session.commit()
-        flash(f'Usuário {user.name} adicionado!')
+        endereco = UserEnderecoModel(
+            user_id=user.id,
+            cep=form.cep.data,
+            numero=form.numero.data,
+            bairro=form.bairro.data,
+            endereco=form.endereco.data,
+            complemento=form.complemento.data
+        )
+        db.session.add(endereco)
+        db.session.commit()
+        form = UserForm(formdata=None)
+        flash(f'User {user.name} added!')
         return redirect(url_for('add_user'))
-    else:
-        return render_template('user.html',form=form)
+
+    return render_template('user.html', form=form)
+
 
 
 @app.route('/alter_user/<int:id>', methods=['POST','GET'])
@@ -69,7 +88,8 @@ def alter_user(id):
         user.email = form.email.data
         user.username = form.username.data
         if foto != None:
-            os.remove(os.path.join(app.config['UPLOAD_FOLDER'], user.profile_pic))
+            if user.profile_pic != None:
+                os.remove(os.path.join(app.config['UPLOAD_FOLDER'], user.profile_pic))
             profile_pic = "{}_{}".format(uuid1(),secure_filename(foto.filename))
             user.profile_pic = profile_pic
             foto.save(os.path.join(app.config['UPLOAD_FOLDER'], profile_pic))
@@ -80,6 +100,23 @@ def alter_user(id):
     else:
         return render_template('alter_user.html',form=form, user=user)
 
+@app.route('/desabilitar_usuario')
+@login_required
+def desativar_usuario():
+    user = UserModel.query.get_or_404(current_user.id)
+    user.status = False
+    db.session.commit()
+    flash(f'{current_user.username} foi desabilitado')
+    return redirect(url_for('dashboard_user'))
+
+@app.route('/habilitar_usuario')
+@login_required
+def ativar_usuario():
+    user = UserModel.query.get_or_404(current_user.id)
+    user.status = True
+    db.session.commit()
+    flash(f'{current_user.username} foi habilitado')
+    return redirect(url_for('dashboard_user'))
 
 @app.route('/alter_user_password/<int:id>', methods=['POST','GET'])
 @login_required
@@ -130,7 +167,7 @@ def add_post():
     else:
         return render_template('user_post.html', form=form)
 
-@app.route('/post/edit/<int:id>',methods=['POST','GET'])
+@app.route('/alterar/<int:id>',methods=['POST','GET'])
 @login_required
 def alter_post(id):
     form = UserPostForm()
@@ -138,8 +175,9 @@ def alter_post(id):
     if post.poster.id == current_user.id:
         if form.validate_on_submit():
             foto = form.prato_pic.data
-            if foto != None:
-                os.remove(os.path.join(app.config['UPLOAD_FOLDER'], post.prato_pic))
+            if foto:
+                if post.prato_pic != None:
+                    os.remove(os.path.join(app.config['UPLOAD_FOLDER'], post.prato_pic))
                 prato_pic = "{}_{}".format(uuid1(),secure_filename(foto.filename))
                 post.prato_pic = prato_pic
                 foto.save(os.path.join(app.config['UPLOAD_FOLDER'], prato_pic))
@@ -149,12 +187,13 @@ def alter_post(id):
             post.preco = form.preco.data
             form = UserPostForm(formdata=None)
             db.session.commit()
-            flash('Post Alterado!')
-            return redirect(url_for('index'))
+            flash(f'{post.name} Alterado!')
+            return redirect(url_for('dashboard_user'))
         else:
             form.name.data = post.name
             form.categoria.data = post.categoria 
             form.content.data = post.content
+            
             form.prato_pic.data = post.prato_pic
             form.preco.data = post.preco
             return render_template('alter_post.html',form=form, post=post)
@@ -162,20 +201,31 @@ def alter_post(id):
         flash('Você não pode alterar esse prato!')
         return redirect(url_for('index'))
 
-@app.route('/post/delete/<int:id>')
+@app.route('/desabilitar/<int:id>')
 @login_required
-def delete_post(id):
-    post = UserPostModel.query.get_or_404(id)
-    id = current_user.id
-    if id == post.poster.id:
-        db.session.delete(post)
+def desabilitar_prato(id):
+    post = UserPostModel.query.get_or_404(id) 
+    if current_user.id == post.poster.id:
+        post.status = False
         db.session.commit()
-        flash(f"Prato deletado!")
-        return redirect(url_for('index'))
+        flash(f"{post.name} desabilitado!")
+        return redirect(url_for('dashboard_user'))
     else:
         flash(f"Você não pode deletar esse prato!")
         return redirect(url_for('index'))
 
+@app.route('/ativar/<int:id>')
+@login_required
+def habilitar_prato(id):
+    post = UserPostModel.query.get_or_404(id) 
+    if current_user.id == post.poster.id:
+        post.status = True
+        db.session.commit()
+        flash(f"{post.name} habilitado!")
+        return redirect(url_for('dashboard_user'))
+    else:
+        flash(f"Você não pode deletar esse prato!")
+        return redirect(url_for('index'))
 
 
 @app.route('/login',methods=['POST','GET'])
@@ -211,14 +261,15 @@ def logout():
 @app.route('/dashboard', methods=['POST','GET'])
 @login_required
 def dashboard_user():
-    user = UserModel.query.filter_by(id=current_user.id).first_or_404()
-    return render_template('dashboard.html',user=user)
+    user = UserModel.query.get_or_404(current_user.id)
+    pratos = UserPostModel.query.filter(UserPostModel.user_id==current_user.id).all()
+    return render_template('dashboard.html',user=user, pratos=pratos)
 
 
 @app.route('/cozinheiros', methods=['POST','GET'])
 def listar_usuarios():
     form = SearchForm()
-    cozinheiros = UserModel.query.all()
+    cozinheiros = UserModel.query.filter(UserModel.status == True).all()
     if form.validate_on_submit():
         search = form.search.data
         cozinheiros = UserModel.query.filter(UserModel.username.like('%{}%'.format(search))).all()
